Factor out broadcast types
