+++ src/voice-broadcast/index.ts
@@ -19,8 +19,7 @@ limitations under the License.
  * {@link https://github.com/vector-im/element-meta/discussions/632}
  */
 
-import { RelationType } from "matrix-js-sdk/src/matrix";
-
+export * from "./types";
 export * from "./models/VoiceBroadcastPlayback";
 export * from "./models/VoiceBroadcastPreRecording";
 export * from "./models/VoiceBroadcastRecording";
@@ -55,26 +54,3 @@ export * from "./utils/startNewVoiceBroadcastRecording";
 export * from "./utils/textForVoiceBroadcastStoppedEvent";
 export * from "./utils/textForVoiceBroadcastStoppedEventWithoutLink";
 export * from "./utils/VoiceBroadcastResumer";
-
-export const VoiceBroadcastInfoEventType = "io.element.voice_broadcast_info";
-export const VoiceBroadcastChunkEventType = "io.element.voice_broadcast_chunk";
-
-export type VoiceBroadcastLiveness = "live" | "not-live" | "grey";
-
-export enum VoiceBroadcastInfoState {
-    Started = "started",
-    Paused = "paused",
-    Resumed = "resumed",
-    Stopped = "stopped",
-}
-
-export interface VoiceBroadcastInfoEventContent {
-    device_id: string;
-    state: VoiceBroadcastInfoState;
-    chunk_length?: number;
-    last_chunk_sequence?: number;
-    ["m.relates_to"]?: {
-        rel_type: RelationType;
-        event_id: string;
-    };
-}
